feat(phone-input): allow consumer to provide countries list
